-Improved/cleanup of Exception reporting;
git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS02]$/alphafs/Development;C111483

+++ AlphaFS/Filesystem/BackupFileStream.cs
@@ -489,7 +489,7 @@ namespace Alphaleonis.Win32.Filesystem
          try
          {
             if (lastError != Win32Errors.ERROR_SUCCESS)
-               NativeError.ThrowException(lastError);
+               NativeError.ThrowException((int) lastError);
 
             if (pSecurityDescriptor.IsInvalid)
                throw new IOException(Resources.InvalidSecurityDescriptorReturnedFromSystem);
